refactor(cypress-screenplay): simplifiy interactions api
Drastically simplifies the API for creating interactions.

BREAKING CHANGE: Use `createTask` and `createQuestion` to create custom interactions instead of
subclassing. Cypress specific implementations moved to `createCypressTask` and
`createCypressInteraction`.
