+++ packages/npm/cypress-screenplay/cypress/support/index.js
@@ -1,1 +1,1 @@
-import '../../dist/commands';
+import '../../dist';
