+++ packages/npm/cypress-screenplay/cypress/support/interactions.js
@@ -1,11 +1,11 @@
-import { createQuestion, createTask } from '../../dist';
+import { createCypressQuestion, createCypressTask } from '../../dist';
 
-export const visitTestPage = createTask((cy, page) => {
-  cy.visit(`./cypress/fixtures/${page}`);
+export const visitTestPage = createCypressTask((cy, page) => {
+  cy.cy.visit(`./cypress/fixtures/${page}`);
 });
 
-export const readListItems = createQuestion((cy, param, assert) => {
-  cy.get('li').should((items) =>
+export const readListItems = createCypressQuestion((cy, _, assert) => {
+  cy.cy.get('li').should((items) =>
     assert(
       items
         .toArray()
