+++ packages/npm/cypress-screenplay/package.json
@@ -11,8 +11,7 @@
     "formatting": "prettier --check \"**/*.{js,jsx,ts,tsx,json,md}\"",
     "formatting:fix": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\"",
     "prepare": "tsc",
-    "pretest": "tsc && yarn lint && yarn formatting ",
-    "test": "mocha -r ts-node/register __tests__/*.spec.ts && cypress run",
+    "test": "tsc --noEmit && yarn lint && yarn formatting && mocha -r ts-node/register src/__tests__/*.spec.ts && cypress run",
     "cypress": "cypress open"
   },
   "devDependencies": {
